Added a complete theme for Ublock Origin
There seems to be a problem with Firefox not loading certain divs for some extensions on startup. In this case it applies to the "#extraTools" and "#saveRules, #revertRules" classes (aka the bottom bar for ublock popup with the "hide cosmetic filters", and the "save rules permanently" buttons). If you check the actual html for the popup page they DO load, so it's not a problem with the code. (you can check it opening the control panel of ublock and then changing the "dashboard.html" last part of the url by "popup.html"