Added a different file for the addons
I decided to split up the addons on a different file (that will be called from usercontent.css) so that it doesn't get as annoying to update usercontent for those that do use any of the addons themed here (since you'd have to update the dynamic urls with each update of the theme if you just copy&paste usercontent).

I fixed the few things left unthemed on ublock and themed other addons such as Flash Video Downloader for youtube, or Video Download Helper (since these are some of the most popular). 

To use this file you'd have to update your usercontent to the new version.